% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPA.R
\name{calc_PPA}
\alias{calc_PPA}
\title{calc_PPA}
\usage{
calc_PPA(STP, t, points = NULL, qs = 12)
}
\arguments{
\item{STP}{A STP_Track}

\item{points}{The points used for the PPA calculation given as a vector of integers.
Default are the space-time points are directlty before and after space-time point}

\item{time}{Time("POSIXct" or"POSIXt") for which the PPA needs to be calculated.}
}
\value{
The Potential Path Area as SpatialPolygons for the provided time
}
\description{
This function calcualtes the Potential Path Area(PPA) for a given moment in time.
}
\author{
Mark ten Vregelaar
}
