% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPA.R
\name{calcPPA_STP}
\alias{calcPPA_STP}
\title{calcPPA_STP}
\usage{
calcPPA_STP(STP, x_density = 250)
}
\arguments{
\item{STP}{A STP_Track}

\item{x_density}{The amount of x coordinates for which the corresponding y coordinate(s) will be calculated.}
}
\value{
The Potential Path Area of the STP as SpatialPolygons
}
\description{
This function calcualtes the Potential Path Area(PPA) of Space-Time Prism(STP).
A STP is a set of space-time points with a maximum speed
}
\author{
Mark ten Vregelaar
}
