% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPA.R
\name{calcPPA_STP_Tinterval}
\alias{calcPPA_STP_Tinterval}
\title{calcPPA_STP_Tinterval}
\usage{
calcPPA_STP_Tinterval(STP_track, time_range, x_density, time_interval, quadsegs)
}
\arguments{
\item{STP_track}{The STP_track for which the PPA needs to be calculated}

\item{time_range}{time range ("POSIXct" or"POSIXt") for which the PPA needs to be calculated.}

\item{x_density}{Paramter used for calculating the PPA of entire STPs.
The amount of x coordinates for which the corresponding y coordinate(s) will be calculated.
Only relevant if the PPA for at least 1 complete STP needs to be calculated}

\item{quadsegs}{Passed to buffer. Number of line segments to use to approximate a quarter circle.
Only used where paramter time_interval is relavant}

\item{time_interval(minutes)}{The time interval used for calculating the PPA.
Only used for calculating the PPA for a specfic moment in time and
if only a part of the PPA of a STP needs to be calculated.
Default is every minute}
}
\value{
The Potential Path Area for the time range as SpatialPolygons
}
\description{
Function for calculating PPA for specific time range
}
\author{
Mark ten Vregelaar
}
