% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPA.R
\name{calcPPA_STP_Track}
\alias{calcPPA_STP_Track}
\title{calcPPA_STP_Track}
\usage{
calcPPA_STP_Track(STP_track, x_density = 250)
}
\arguments{
\item{STP_track}{The STP_Track for which the PPA will be calculated}

\item{x_density}{The amount of x coordinates for which the corresponding y coordinate(s) will be calculated.}
}
\value{
The Potential Path Area of the STP as SpatialPolygons
}
\description{
This function calcualtes the Potential Path Area(PPA) of an entire STP_track
}
\author{
Mark ten Vregelaar
}
