% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPA.R
\name{chull_poly}
\alias{chull_poly}
\title{Function that calculates PPA based on convex hull of PPAcoords}
\usage{
chull_poly(PPAcoords, crs)
}
\description{
Arg:
         PPAcoords(matrix): x and y coordiantes
         crs(CRS): crs of coordinates
}
\details{
Return:
        PPA(SpatialPolygons): Potential Path Area
}
