% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{getMinimalSpeed}
\alias{getMinimalSpeed}
\title{getMinimalSpeed}
\usage{
getMinimalSpeed(track)
}
\arguments{
\item{track}{the trajectory as \link{STP_Track} or \link{Track}}
}
\value{
Vector of minimal speeds in unit spatial projection unit/s The speed required to reach the next point
in the available time
}
\description{
This functions calculates the minimal speed required to reach every point.
The  speed is based on linear movement between two points.
}
\author{
Mark ten Vregelaar
}
