% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{getVmaxtrack}
\alias{getVmaxtrack}
\title{getVmaxtrack}
\usage{
getVmaxtrack(track)
}
\arguments{
\item{track}{the trajectory as \link{STP_Track} or \link{Track}}
}
\value{
max speed of the moving object
}
\description{
This functions calculates the maximum speed found in a trajectory.
The maximum speed is based on linear movement between measured points.
}
\examples{
## get a track dataset
data(A3)# see trajectories package: importenvirocar
car_track<-Track(A3)# recalculate connections data
head(car_track@connections)# distance in meters and speed in m/s
## get maximum speed
speed_ms <- getVmaxtrack(car_track)# speed in m/s
# maximum speed. It is the speed required to be able to reach every point in time in km/h
speed_ms*3.6
}
\author{
Mark ten Vregelaar
}
