% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPA.R
\name{in_time_range}
\alias{in_time_range}
\title{Function that check if the time is within the time_range
   Arg:
         time (POSIXct): time
         time_range(two POSIXct): the time range}
\usage{
in_time_range(time, time_range)
}
\description{
Return:
         True or False(logical): True if time within time_Range
}
