% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squeared}
}
\value{
The square of the input
}
\description{
Take in any numeriv value and squares it.
}
