% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STPtrajectories.R
\docType{package}
\name{STPtrajectories}
\alias{STPtrajectories}
\alias{STPtrajectories-package}
\title{STPtrajectories.}
\description{
Package that uses Space-Time Prism (STP) concept to analyse trajectories.
By incorparting the STP concept, users can analyse their data while taking into account
the uncertainty of the location of an individual in between the control points of a trajectory.
The package also supports uncertainty about the control points themselves.
The STPtrajectories package contains methods that calculate Potential Path Areas(PPAs), create random
trajectories,calculate when an individual could have been at a spatial location, and test for possible encounters between two individuals by applying the alibi query.
It also provides a method that visulises STP trajectories in 3D.
}
\section{Background}{

A trajectory consists of successive points in space and time.
The location of the individual between two successive points is unknown,
but based on a maximum speed a space-time prism can be calculated.
A space-time prism(STP) is defined as the collection of space-time locations the individual could reach,
given a speed limitation. The \link{STP_Track} class can be used to handle space-time prism trajectories.

The \link{alibi_query} uses this concept to test if two individuals could have met each other.
The package provides related functions to help users in the analysis
 of their trajectories and take into account the uncertainty about the location of an individual.
These include methods to visualise space-time prisms(\link{STP_plot}),
calculate the potential path area(\link{PPA}), calculate when an individual could have been at a spatial location(\link{potential_stay})
, and a random trajectory generator(\link{RTG})
}

\section{help}{

 \strong{need help:}
\itemize{
\item manual: \url{https://github.com/markvregel/STPtrajectories/blob/master/STPtrajectories.pdf}
\item vignette: \url{http://htmlpreview.github.io/?https://raw.githubusercontent.com/markvregel/STPtrajectories/master/vignettes/STP_Tracks.html}
\item check functions help and examples
}
}

\seealso{
\itemize{
\item github: \url{https://github.com/markvregel/STPtrajectories}
\item trajecotries package \url{https://cran.rstudio.com/web/packages/trajectories/index.html}
}
}
\author{
Mark ten Vregelaar
}
