% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alibi_query.R
\name{alibi_query}
\alias{alibi_query}
\title{alibi_query}
\usage{
alibi_query(STP_track1, STP_track2)
}
\arguments{
\item{STP_track1}{STP_track1}

\item{STP_track2}{STP_track2}
}
\value{
True or False for the alibi query
}
\description{
CONTAINS ERRORS. CASE 3 goed wrong!!!
This function tests wether there was a possible meeting between two individuals or other moving objects.
If the individuals of two trajecories could have met is tested by applying the alibi query to segments that overlap in time.
The alibi query is a Boolean query that checks whether two moving individuals,
that are given by two samples of space-time points and speed limitations, could have met each other.
The query tests if two space-time prisms intersect.
Kuijpers et al. (2011) provide the analytical solution for the alibi query that is used by this function
}
\references{
- 	Kuijpers, B., Grimson, R., & Othman, W. (2011).
An analytic solution to the alibi query in the space-time prisms model for moving object data.
International Journal of Geographical Information Science, 25(2), 293-322.
}
\author{
Mark ten Vregelaar
}
