% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{getVmaxtrack}
\alias{getVmaxtrack}
\title{getVmaxtrack}
\usage{
getVmaxtrack(track)
}
\arguments{
\item{track}{the trajectory as \link{STP_Track} or \link{Track}}
}
\value{
max speed of the moving object
}
\description{
This functions calculates the maximum speed found in a trajectory.
The maximum speed is based on linear movement between measured points.
}
\author{
Mark ten Vregelaar
}
