% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STP_Track.R
\name{STP_Track}
\alias{STP_Track}
\title{STP_Track class}
\usage{
STP_Track(track, vmax)
}
\arguments{
\item{Track}{object of class  \link{Track}}
}
\description{
A class to represent Space-Time Prism(STP) trajecories.
These are trajectories with a maximum speed for each segment.
The maximum speed is added to the connections slot
}
\section{Slots of class "Track"}{

  \describe{
	\item{\code{sp}:}{spatial locations of the track points, with length n}
	\item{\code{time}:}{time stamps of the track points}
	\item{\code{endTime}:}{end time stamps of the track points}
	\item{\code{data}:}{\code{data.frame} with n rows, containing
	attributes of the track points}
    \item{\code{connections}:}{\code{data.frame}, with n-1 rows, containing
	attributes between the track points such as distance and speed }
  }

}

